requirements.txt
pymupdf==1.26.3
qdrant-client==1.15.1
fastembed==0.7.1
streamlit==1.48.0
python-dotenv==1.1.1
asyncio==4.0.0
pandas==2.3.1
numpy==2.3.2
sentence-transformers==2.2.2
rank-bm25==0.2.2
langchain==0.1.0
langchain-google-genai==0.0.6
sqlite3